+++ fedlab_benchmarks/algorithm/fedavg/hierarchical_demo/run.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+python client_hi.py &
+python server_hi.py &
+python middle.py &
